+++ lib/capybara_accessible_selectors/selectors/dialog.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Capybara.add_selector(:dialog, locator_type: [String, Symbol]) do
+Capybara.add_selector(:dialog, locator_type: [String, Regexp]) do
   xpath do |*|
     XPath.descendant[
       [
@@ -10,14 +10,13 @@ Capybara.add_selector(:dialog, locator_type: [String, Symbol]) do
     ]
   end
 
-  locator_filter do |node, locator, exact:, **|
-    next true if locator.nil?
-
-    method = exact ? :eql? : :include?
-    if node[:"aria-labelledby"]
-      CapybaraAccessibleSelectors::Helpers.element_labelledby(node).public_send(method, locator)
-    elsif node[:"aria-label"]
-      node[:"aria-label"].public_send(method, locator.to_s)
+  locator_filter skip_if: nil do |node, locator, exact:, **|
+    accessible_name = node.accessible_name
+    case locator
+    when String
+      exact ? accessible_name == locator : accessible_name.include?(locator.to_s)
+    when Regexp
+      locator.match?(accessible_name)
     end
   end